Create mersenne_sequence_generator.php
 first check if the request method is POST and if the 'n' parameter is provided. If not, we return appropriate error messages.

Then, we retrieve the value of 'n' from the request and validate it. If it's a valid positive integer, we call the generateMersenneSequence() function to generate the Mersenne sequence.

The generateMersenneSequence() function generates the Mersenne sequence of length 'n' by checking if each number (starting from 2) is a Mersenne prime using the isPrime() helper function. If it is, we add it to the sequence until we reach the desired length 'n'.

The isPrime() function checks if a given number is prime by iterating from 2 to the square root of the number and checking for any divisors. If a divisor is found, the number is not prime.

Finally, we encode the generated sequence as JSON and send it back as the response to the client-side code.